Fix detection with OpenSSH 7.6
Right now, we look only at lines with "sign_and_send_pubkey", but in
some OpenSSH versions, such as 7.6, those lines don't indicate whether
SHA-2 is being used.  We can fix this by looking at the "Server accepts
key" line, which contains the explicit algorithm name.

Note that it's technically possible for the server to accept the key and
then refuse the signature if the signature is bad, but we know OpenSSH
will not produce such a signature and GitHub will not otherwise refuse
such a signature.  In addition, PuTTY will scream very loudly if this
happens, so it's not likely that other servers will do so, either.